Fixes #94: Accessing app_label and model_name through content type rather than object.
As the object property is not always loaded by prefetch_related(), the `search_result_item` template tag will raise an AttributeError. But, as the object is essentially used for populating the string format params dict, it is possible to instead use the content type object (which appears to be available even when object is missing) for getting the `app_label` and `model_name` strings.

The only drawback is that the `obj` template variable may be `None` in some instances. It is not accessed in any standard tempaltes, and this issue would have caused an AttributeError earlier anyway.